porcofile

+++ client.c
@@ -1,10 +1,11 @@
-#include <signal.h>
-#include <stdio.h>
-#include <unistd.h>
-#include <stdlib.h>
+#include "minitalk.h"
 
 int main(int ac, char **av)
 {
-	int pid = atoi(av[1]);
-	kill(pid, SIGUSR1);
+	if (ac == 2)
+	{
+		int pid;
+		pid = ft_atoi(av[1]);
+		kill(pid, SIGUSR1);
+	}
 }
